backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.26.0
aiofiles>=23.2.1
